+++ src/clojure/clojurewerkz/welle/ring/session_store.clj
@@ -11,7 +11,8 @@
   (:require [ring.middleware.session.store :as ringstore]
             [clojurewerkz.welle.kv         :as kv])
   (:import [java.util UUID Date]
-           com.basho.riak.client.http.util.Constants))
+           com.basho.riak.client.http.util.Constants
+           com.basho.riak.client.raw.RawClient))
 
 ;;
 ;; Implementation
@@ -32,14 +33,14 @@
 ;; API
 ;;
 
-(defrecord RiakSessionStore [^String bucket-name r w content-type])
+(defrecord RiakSessionStore [^RawClient client ^String bucket-name r w content-type])
 
 (extend-protocol ringstore/SessionStore
   RiakSessionStore
 
   (read-session [store key]
     (if-let [m (and key
-                    (let [reply (kv/fetch-one (.bucket-name store) key :r (.r store))
+                    (let [reply (kv/fetch-one (.client store) (.bucket-name store) key {:r (.r store)})
                           result (:result reply)]
                       (when result
                         (:value result))))]
@@ -48,23 +49,23 @@
 
   (write-session [store key data]
     (let [key (or key (str (UUID/randomUUID)))]
-      (kv/store (.bucket-name store) key (assoc data :date (Date.))
-             :content-type (.content-type store) :w (.w store))
+      (kv/store (.client store) (.bucket-name store) key (assoc data :date (Date.))
+                {:content-type (.content-type store) :w (.w store)})
       key))
 
   (delete-session [store key]
-    (kv/delete (.bucket-name store) key :w (.w store))
+    (kv/delete (.client store) (.bucket-name store) key {:w (.w store)})
     nil))
 
 
 (defn welle-store
-  ([]
-     (RiakSessionStore. default-session-store-bucket default-r default-w Constants/CTYPE_JSON_UTF8))
-  ([^String bucket-name]
-     (RiakSessionStore. bucket-name default-r default-w Constants/CTYPE_JSON_UTF8))
-  ([^String bucket-name r]
-     (RiakSessionStore. bucket-name r default-w Constants/CTYPE_JSON_UTF8))
-  ([^String bucket-name r w]
-     (RiakSessionStore. bucket-name r w Constants/CTYPE_JSON_UTF8))
-  ([^String bucket-name r w ^String content-type]
-     (RiakSessionStore. bucket-name r w content-type)))
+  ([^RawClient client]
+     (RiakSessionStore. client default-session-store-bucket default-r default-w Constants/CTYPE_JSON_UTF8))
+  ([^RawClient client ^String bucket-name]
+     (RiakSessionStore. client bucket-name default-r default-w Constants/CTYPE_JSON_UTF8))
+  ([^RawClient client ^String bucket-name r]
+     (RiakSessionStore. client bucket-name r default-w Constants/CTYPE_JSON_UTF8))
+  ([^RawClient client ^String bucket-name r w]
+     (RiakSessionStore. client bucket-name r w Constants/CTYPE_JSON_UTF8))
+  ([^RawClient client ^String bucket-name r w ^String content-type]
+     (RiakSessionStore. client bucket-name r w content-type)))