+++ src/clojure/clojurewerkz/welle/testkit.clj
@@ -7,11 +7,12 @@
 ;; the terms of this license.
 ;; You must not remove this notice, or any other, from this software.
 
-(ns ^{:doc "Utility functions useful for unit and integration testing of applications
-            that use Welle"}
-  clojurewerkz.welle.testkit
+(ns clojurewerkz.welle.testkit
+  "Utility functions useful for unit and integration testing of applications
+   that use Welle"
   (:require [clojurewerkz.welle.kv      :as kv]
-            [clojurewerkz.welle.buckets :as wb]))
+            [clojurewerkz.welle.buckets :as wb])
+  (:import com.basho.riak.client.raw.RawClient))
 
 ;;
 ;; API
@@ -21,6 +22,7 @@
   "Drains the bucket with the provided name by deleting all the keys in it. For buckets with
    a large number of keys this may be a very expensive operation because it involves listing keys
    in the bucket."
-  [^String bucket-name]
-  (doseq [k (wb/keys-in bucket-name)]
-    (kv/delete bucket-name k :w 1)))
+  [^RawClient client ^String bucket-name]
+  (doseq [k (wb/keys-in client bucket-name)]
+    (kv/delete client bucket-name k {:w 1})))
+