+++ test/clojurewerkz/welle/test/indices_pb_test.clj
@@ -8,19 +8,18 @@
   (:import java.util.UUID
            com.basho.riak.client.http.util.Constants))
 
-(wc/connect-via-pb!)
-
 (deftest ^{:2i true :edge-features true} test-basic-index-query-with-a-single-string-value-over-pb
-  (let [bucket-name "clojurewerkz.welle.test.indices-pb-test"
-        bucket      (wb/update bucket-name)
+  (let [conn        (wc/connect-via-pb)
+        bucket-name "clojurewerkz.welle.test.indices-pb-test"
+        bucket      (wb/update conn bucket-name)
         k           (str (UUID/randomUUID))
         v           (.getBytes "value")
         indexes     {:email #{"johndoe@example.com" "timsmith@example.com"}}
-        stored      (kv/store bucket-name k v :indexes indexes :content-type Constants/CTYPE_OCTET_STREAM)
-        [idx-key]   (kv/index-query bucket-name :email "johndoe@example.com")
-        [fetched]   (kv/fetch bucket-name idx-key)]
+        stored      (kv/store conn bucket-name k v {:indexes indexes :content-type Constants/CTYPE_OCTET_STREAM})
+        [idx-key]   (kv/index-query conn bucket-name :email "johndoe@example.com")
+        [fetched]   (:result (kv/fetch conn bucket-name idx-key))]
     (is (:indexes fetched))
     (is (= (String. ^bytes (:value fetched))
            (String. ^bytes v)))
     (is (= (:indexes fetched) indexes))
-    (kv/delete bucket-name k)))
+    (kv/delete conn bucket-name k)))